Had an issue previously trying to upload this; ran into Git giant file issue. Reset master back to last working commit (Peter's vector module) and am now re-pushing this without the big files. I'm hoping that removing those from the repo won't ruin the ability to run the project in Unity but the executable is working as a WebGL file.
Mostly working project. It provides almost all of the original intended functionality - just missing file loading.
Instead for the sake of providing a more complete set of documentation and testing; I've focused on just providing the executable for one scene.

I've realised in creating it this way that most of my tests are System tests as they're about checking the output as controlled by UI. I will need to change several tests from the System VnV as this is no longer a plugin to be installed but an executable that can be run. This change was made to make it easier for other students to use and compare how lighting models are affected by the elements covered in my Commonality Analysis. I think this showcases the problem better than a plugin would have.

To do:
------
1. Implement code for Sharon
2. Finish testing
3. Revise documents